// management/src/routes/media/__generated__/MediaLibraryFragment.graphql.ts
export type MediaLibraryFragment$data = {
  readonly medias: {
    readonly edges: ReadonlyArray<{
      readonly cursor: string;
      readonly node: {
        readonly altText: string | null | undefined;
        readonly caption: string | null | undefined;
        readonly createdAt: any;
        readonly description: string | null | undefined;
        readonly fileName: string;
        readonly fileSizeBytes: number | null | undefined;
        readonly id: string;
        readonly mimeType: string;
        readonly title: string | null | undefined;
        readonly updatedAt: any;
        readonly uploadedByID: string | null | undefined;
        readonly url: string;
      } | null | undefined;
    }>;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
    readonly totalCount: number;
  };
  readonly " $fragmentType": "MediaLibraryFragment";
};